feat: 로또 지불 기능 추가
로또 지불 기능 추가
- 로또 지불을 domain으로 이동
- 로또 지불에 잔액 정보 저장
- 입력한 개수만큼 로또 구매
- 남은 잔액 확인
